migrate to jpa/postgresSQL database

+++ src/main/java/com/example/fantasticrestaurantv1/restaurants/repository/RestaurantItemRepository.java
@@ -9,5 +9,5 @@ public interface RestaurantItemRepository extends JpaRepository<RestaurantItem, 
 
     RestaurantItem findRestaurantItemByName(String name);
 
-    Collection<RestaurantItem> findAllRestaurantItemsById(Integer id);
+    Collection<RestaurantItem> findAllRestaurantItemsById(Long id);
 }
